Remove tons of async chunk stuff. Fixed invisible chunk bug
Loading of chunks on movement is now waited on in main thread. Also population of a chunk, generating trees and ore fields, is part of the loading process now. This simplifies the code of chunk sending a ton making it faster and less error prone. The only downside is for new generated chunks, since the main thread waits until all chunks are fully loaded, it may lag the main thread on huge view distances.